pyproject.toml
[project]
name = "oraculo-bot"
version = "0.1.0"
description = "Bot Discord de estudos para concursos públicos brasileiros"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "agno>=1.0.0",
    "discord.py>=2.4.0",
    "requests>=2.32.0",
    "openai>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.2.0",
    "supabase>=2.28.3",
    "google-genai>=1.0.0",  # Para embeddings Gemini (free-tier) - pacote novo
]

[project.scripts]
oraculo-bot = "oraculo_bot.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["oraculo_bot"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
